Add a responsive navbar to the webpage
I created a navigation bar within anchor tags, imitating a fully functional webpage. For the layout and alignment, I used the flexbox model. I also added hover effects to improve interaction by users.

There is a border-top which I added to separate it from the h1 however this was just for the purpose of design and making the navigation bar look more pronounced.