Learn about CSS static and relative positioning
I created a simple webpage using the a flexbox display with an article and aside bar. I then implemented static and relative positions for CSS onto this page.

I learnt that the static position is the default which explains why people mainly use it when changing the display on things like media queries where the position has been changed.

I also learnt about how the relative position allows you to move an item throughout the page without affecting the structure, and it can also be overlapped by any items/text which it might hover over.