Completed the layout for all resolutions using media queries
In this commit, I complete the last type of resolution which would typically be for tablets.

An issue I managed to solve was that I did try to increase the font size of the article and I was adding the style to the main element which caused no change in the font size however I quickly solved this by realising my mistake.

An issue I haven't figured out how to solve yet is centring the nav bar horizontally on the phone and tablet resolutions.

There is also an issue which I faced when I was styling the entire page without media queries and it is that a gap appears between the article and aside elements when the code is rendered if the footer is taking up the full width on the page whilst using the grid layout, however when I make it take up only 80% of the page's width then there no longer appears the gap between the article and aside elements.